+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## 2.14.0 (????-??-??)
 ### New Modules
-- State Variable Filter
+- Multi Mode Filter
 
 ## 2.13.2 (2025-10-05)
 ### New Modules